permissions when starting slyd and nginx
When building the vagrant slyd and nginx is not correctly started because it lacks permissions to access logs etc and you should "vagrant ssh" and start these services manually. I have added sudo to both commands and changed "nginx start" for "nginx restart". This way i get it to start correctly.

I found this problem with this setup:

* OsX 10.11.2 El Capitán
* vagrant 1.8.1