Fix splash browser tab storing wrong html
If a user loaded a page, Then loaded a different page, then went back to the
initial page, then created a sample the html stored in the `original_body.html`
file may have been incorrect. This is caused by the html being stored during
`reply_read` phase of the network manager which is not called on a subsequent
page load due to caching. The solution used here is just to reload the page
if the html is incorrect, this is found through comparing the url in the frame
and in the network manager

Other Changes:
Send page metadata to the UI on page load finished
Do no evaljs or load page url if the tab is in a closing state